♻️Refactor: 자산 목록 API 통신 모듈화 #94
- 마이페이지-입출금, 메인-주문창에서 사용
