+++ view/src/widgets/CashTransaction/index.tsx
@@ -1,22 +1,19 @@
 import { useState } from 'react';
 import MyAssetInfo from '../../entities/MyAssetInfo';
 import MyAssetList from '../../entities/MyAssetList';
-import useGetAssets from './model/useGetAssets';
+import useGetAssets from '../../shared/model/useGetAssets';
 
 const CashTransaction = () => {
   const [selectedAssetIdx, setSelectedAssetIdx] = useState(0);
 
-  const { data: assetList } = useGetAssets();
+  const { data: assets } = useGetAssets();
 
-  return assetList ? (
+  return assets ? (
     <div>
-      <MyAssetList
-        assetList={assetList.assets}
-        setSelectedAssetIdx={setSelectedAssetIdx}
-      ></MyAssetList>
+      <MyAssetList assetList={assets} setSelectedAssetIdx={setSelectedAssetIdx}></MyAssetList>
       <MyAssetInfo
-        currencyCode={assetList.assets[selectedAssetIdx].currencyCode}
-        amount={assetList.assets[selectedAssetIdx].amount}
+        currencyCode={assets[selectedAssetIdx].currencyCode}
+        amount={assets[selectedAssetIdx].amount}
       ></MyAssetInfo>
     </div>
   ) : (
