--- view/src/widgets/CashTransaction/model/useGetAssets.ts
@@ -1,12 +1,0 @@
-import { useQuery } from '@tanstack/react-query';
-import getAssetsApi from '../../../widgets/CashTransaction/api/getAssetsApi';
-
-const useGetAssets = () => {
-  return useQuery({
-    queryKey: ['assets'],
-    queryFn: getAssetsApi,
-    retry: 1,
-  });
-};
-
-export default useGetAssets;
